Move tokens config under `plugins` key
Updating the tokens configuration to be aligned with the format proposed
in `FIR-12: Namepsace Isolation`. The config is now under a top-level key called `plugins`,
with a slightly updated format. The existing config is now marked as
deprecated, but still supported.

There is not support for having both configration structures defined, the new
`plugins` config structure takes precedence over the deprecated one.

Example of the new configuration:
```
plugins:
  tokens:
      - type: fftokens
        name: erc20_erc721
        fftokens:
          url: http://tokens_0_0:3000
```

Signed-off-by: Alex Shorsher <alex.shorsher@kaleido.io>