There Can Only Be One [multiparty V1 contract]
Because ff_system is global, the contract instance is global. The address
can never change, and a second V1 contract can never be used. Adjust the
core logic to enforce this.

This also means splitting the E2E suite, as a V2->V2 migration is simple
to test over and over, but a V1->V2 migration can only be done once per
stack. This special test will have to be run infrequently under controlled
circumstances.

Signed-off-by: Andrew Richardson <andrew.richardson@kaleido.io>