Separate NextPin entries by namespace
Each local node (and therefore local namespace) needs to maintain its own NextPin table.
Previously this was not a problem, since the Context is a function of (among other things)
the remote namespace. Now, in a multi-tenant situation, two local namespaces might be
maintaining separate snapshots of the same remote namespace.

While it's not possible to populate namespace on any existing entries, we make a best
effort at migration by setting the namespace every time a row is updated. This will allow
legacy environments to migrate over time, as long as they don't immediately add
multi-tenant nodes to their setup.

Signed-off-by: Andrew Richardson <andrew.richardson@kaleido.io>