Replace msg.sender with tx.origin in Firefly contract events
We actually want to emit the author that signed the transaction, not the
immediate sender that invoked pinBatch (which might be another smart
contract).

This does imply that deployers of this contract must carefully consider
whether this opens the possibility of other contracts impersonating an
entity on the network by calling "pinBatch" as a side-effect of some
unrelated transaction. Extending this contract with an allow-list of
trusted senders may be prudent in some situations.

Signed-off-by: Andrew Richardson <andrew.richardson@kaleido.io>